+++ concert_software_farmer/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package concert_software_farmer
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.6.9 (2015-07-09)
+------------------
+
 0.6.8 (2015-04-27)
 ------------------
 * delete timeout exception and fix wrong timeout time