+++ rocon_concert/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package rocon_concert
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.6.9 (2015-07-09)
+------------------
+
 0.6.8 (2015-04-27)
 ------------------
 * add concert_software_farmer is missing dependency clsoses `#289 <https://github.com/robotics-in-concert/rocon_concert/issues/289>`_